// src/app/material-type.enum.ts
export enum MaterialType {
    LINE_BASIC = "Line Basic",
    LINE_DASHED = "Line Dashed",
    MESH_BASIC = "Mesh Basic",
    MESH_DEPTH = "Mesh Depth",
    MESH_DISTANCE = "Mesh Distance",
    MESH_LAMBERT = "Mesh Lambert",
    MESH_MATCAP = "Mesh Matcap",
    MESH_NORMAL = "Mesh Normal",
    MESH_PHONG = "Mesh Phong",
    MESH_PHISICAL = "Mesh Phisical",
    MESH_STANDARD = "Mesh Standard",
    MESH_TOON = "Mesh Toon",
    POINTS = "Mesh Points",
    SHADOW = "Shadow",
    SPRITE = "Sprite"
}